+++ app/Http/Controllers/ShopController.php
@@ -42,12 +42,6 @@ class ShopController extends Controller
         return redirect('user/shop/'.$user->id.'/index');
     }
 
-    public function show($id)
-    {
-        $shop = Shop::find($id);
-        return view('shop.show')->with('shop', $shop);
-    }
-
     public function edit($id)
     {
         $shop = Shop::find($id);
@@ -78,39 +72,5 @@ class ShopController extends Controller
         $shop = Shop::find($id);
         $shop->delete();
         return redirect('user/shop/'.$shop->user_id.'/index');
-    }
-
-    public function post($id)
-    {
-        $shop = Shop::find($id);
-        $products = Product::All();
-        return view('shop.post')->with('shop', $shop)->with('products', $products);
-    }
-
-    public function createProduct($id_shop, $id_product)
-    {
-        $shop = Shop::find($id_shop);
-        $product = Product::find($id_product);
-        return view('shop.postProduct')->with('shop', $shop)->with('product', $product);
-    }
-
-    public function postProduct($id, Request $request)
-    {
-        $this->validate($request, [
-            'price' => 'required|numeric',
-            'quantity' => 'required|numeric',
-            'discount' => 'required|numeric',
-        ], [
-            'price.required' => 'Bắt buộc phải điền giá thành',
-            'price.numeric' => 'Giá phải là dạng số',
-            'quantity.required' => 'Bắt buộc phải điền số lượng',
-            'quantity.numeric' => 'Số lượng phải là dạng số',
-            'discount.required' => 'Bắt buộc phải điền giảm giá',
-            'discount.numeric' => 'Giảm giá phải là dạng số',
-        ]);
-        $shop = Shop::find($id);
-        $shopProduct = ShopProduct::find($id);
-        $shopProduct = ShopProduct::create(Input::all());
-        return redirect('user/shop/'.$shop->id.'/show');
     }
 }
