+++ resources/views/shop/index.blade.php
@@ -39,7 +39,7 @@
                                     <td class="text-center">{{$shop->description}}</td>
                                     <td class="text-center">{{$shop->address}}</td>
                                     <td class="text-center">
-                                        <a href="{{ url('user/shop/'.$shop->id.'/show') }}" class="btn btn-info btn-sm pull-right"><span class="glyphicon glyphicon-book"></span> Quản lý shop </a>
+                                        <a href="{{ url('shop/'.$shop->id.'/show') }}" class="btn btn-info btn-sm pull-right"><span class="glyphicon glyphicon-book"></span> Quản lý shop </a>
                                         <a href="{{ url('user/shop/'.$shop->id.'/edit') }}" class="btn btn-success btn-sm pull-right"><span class="glyphicon glyphicon-pencil"></span> Chỉnh sửa </a>
                                         <a href="{{ url('user/shop/'.$shop->id.'/delete') }}" class="btn btn-sm btn-danger pull-right"><span class="glyphicon glyphicon-trash"></span> Xóa </a>
                                     </td>
