+++ resources/views/shop/post.blade.php
@@ -44,7 +44,7 @@
                                     <td class="text-center">{{$product->description}}</td>
                                     <td class="center">{{$product->price}}</td>
                                     <td class="text-center">
-                                        <a href="{{ url('user/shop/post/'.$shop->id.'/'.$product->id.'/create') }}" class="btn btn-success btn-sm"><span class="glyphicon glyphicon-ok"></span> Chọn sản phẩm </a>
+                                        <a href="{{ url('shop/post/'.$shop->id.'/'.$product->id.'/create') }}" class="btn btn-success btn-sm"><span class="glyphicon glyphicon-ok"></span> Chọn sản phẩm </a>
                                     </td>
                                 </tr>
                             @endforeach
